+++ externals/kyuubi-spark-sql-engine/pom.xml
@@ -63,8 +63,13 @@
             <artifactId>spark-hive-thriftserver_${scala.binary.version}</artifactId>
             <scope>test</scope>
         </dependency>
-    </dependencies>
 
+        <dependency>
+            <groupId>org.apache.iceberg</groupId>
+            <artifactId>${iceberg.name}</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 
     <build>
         <outputDirectory>target/scala-${scala.binary.version}/classes</outputDirectory>
